#5 move components
